requirements.txt
flask
torch
requests
python-dotenv
supabase
openai==0.27.7
transformers>=4.32.0
optimum>=1.12.0
auto-gptq --extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/